Move TxRow
