+++ ui/DesignSystem/Component/index.js
@@ -13,7 +13,7 @@ import SidebarLayout from "./SidebarLayout.svelte";
 import SourceBrowser from "./SourceBrowser.svelte";
 import StepCounter from "./StepCounter.svelte";
 import Transaction from "./Transaction.svelte";
-import TxRow from "./TxRow.svelte";
+import TxRow from "./Transaction/Row.svelte";
 import UserCard from "./UserCard.svelte";
 
 export {
